Migrate unpaywall resolver to new structure
- Move UnpaywallResolver from akd/tools/scrapers/resolvers.py to akd/tools/resolvers/unpaywall.py
- Adapt to current BaseArticleResolver structure:
  * Update method signature: resolve(params: ResolverInputSchema) -> ResolverOutputSchema | None
  * Use ResolverOutputSchema with metadata preservation via params.model_dump()
  * Set resolved_url field instead of returning URL string
  * Return None on failure instead of original URL
  * Use async httpx instead of requests
  * Track resolver usage in resolvers list
- Add UnpaywallResolver to akd/tools/resolvers/__init__.py exports
- Update tests to new structure and import paths:
  * Test ResolverOutputSchema return type instead of string
  * Test resolved_url field and resolvers list
  * Verify None returns on resolution failure
  * Update imports from akd.tools.resolvers
